+++ api/src/main/java/ca/bc/gov/educ/api/graduation/config/EducGraduationApplicationConfig.java
@@ -1,0 +1,24 @@
+package ca.bc.gov.educ.api.graduation.config;
+
+import org.modelmapper.ModelMapper;
+import org.springframework.boot.web.client.RestTemplateBuilder;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestTemplate;
+
+@Configuration
+public class EducGraduationApplicationConfig {
+
+    @Bean
+    public ModelMapper modelMapper() {
+
+        ModelMapper modelMapper = new ModelMapper();
+        return modelMapper;
+    }
+
+    @Bean
+    public RestTemplate restTemplate(RestTemplateBuilder builder) {
+        return builder.build();
+    }
+
+}
